+++ modules/core/src/layout/layered/layeredLayout.ts
@@ -5,7 +5,6 @@ import {Point, TriangleOrientation} from '../../math/geometry/point'
 import {Algorithm} from '../../utils/algorithm'
 import {PolyIntEdge} from './polyIntEdge'
 import {SugiyamaLayoutSettings, SnapToGridByY} from './sugiyamaLayoutSettings'
-import {CommonLayoutSettings} from '../commonLayoutSettings'
 
 import {IEdge} from '../../structs/iedge'
 import {CycleRemoval} from './CycleRemoval'
@@ -13,7 +12,7 @@ import {GeomNode} from '../core/geomNode'
 import {Database} from './Database'
 import {LayerArrays} from './LayerArrays'
 import {GeomEdge} from '../core/geomEdge'
-import {edgesAreAttached, GeomGraph, optimalPackingRunner} from '../core/geomGraph'
+import {GeomGraph, optimalPackingRunner} from '../core/geomGraph'
 import {IntPairMap} from '../../utils/IntPairMap'
 import {IntPairSet} from '../../utils/IntPairSet'
 import {IntPair} from '../../utils/IntPair'
